+++ lib/player_engine.rb
@@ -1,2 +1,8 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
+
+class Player
+  def initialize(name)
+    @name = name
+  end
+end
